+++ CHANGELOG.md
@@ -13,7 +13,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### Fixes
 - skip destroy of managed-project-policy if it has roles attached
-
+- if managed-project-policy is in an `*_IN_PROCESS` state, wait 60 seconds and check again
+  - bumps `aws-codeseeder~=0.10.2`
 
 ## v2.10.4 (2023-10-23)